Add support for large messages
- Fix implicit type conversion in reader that caused byte overflow when number of fields in definition message in >85.
- Add real life fit activity to the test quite to replicate the problem
- Fix definition message creation logic in `writer` -- now it scans all records of specific type to find all fields that have valid values to build a definition message.

Fixes #43 and #53

+++ reader.go
@@ -434,7 +434,7 @@ func (d *decoder) parseDefinitionMessage(recordHeader byte) (*defmsg, error) {
 		return &dm, nil
 	}
 
-	if err = d.readFull(d.tmp[0 : 3*dm.fields]); err != nil {
+	if err = d.readFull(d.tmp[0 : 3*uint16(dm.fields)]); err != nil {
 		return nil, fmt.Errorf("error parsing fields: %v", err)
 	}
